requirements.txt
asgiref==3.4.1
autopep8==1.6.0
backports.entry-points-selectable==1.1.1
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.9
cryptography==35.0.0
distlib==0.3.3
Django==3.2.8
django-filter==21.1
django-rest-knox==4.1.0
djangorestframework==3.12.4
filelock==3.4.0
idna==3.3
mysql==0.0.3
mysql-connector-python==8.0.26
mysqlclient==2.1.0
numpy==1.21.4
Pillow==8.4.0
platformdirs==2.4.0
playsound==1.2.2
protobuf==3.18.1
PyAudio @ file:///C:/Users/lokes/Downloads/PyAudio-0.2.11-cp310-cp310-win_amd64.whl
pycodestyle==2.8.0
pycparser==2.21
pycryptodome==3.11.0
pydub==0.25.1
pygame==2.1.0
pynput==1.7.5
pytz==2021.3
requests==2.26.0
scipy==1.7.3
six==1.16.0
sounddevice==0.4.3
SoundFile==0.10.3.post1
sqlparse==0.4.2
toml==0.10.2
urllib3==1.26.7
virtualenv==20.10.0
Wave==0.0.2
